Add solid color themes
This commit introduces solid black and solid white color schemes as fallbacks when dynamic theming is not enabled.

The `SolidDarkColorScheme` uses black for primary, secondary, background, and surface colors, with white for text and icons.
The `SolidLightColorScheme` uses white for primary, secondary, background, and surface colors, with black for text and icons.

The `XArchiverTheme` composable now checks if `dynamicColor` is false and applies these solid color schemes accordingly for dark and light themes.

Signed-off-by: gustyxpower <gustiadityamuzaky08@gmail.com>